+++ app/Http/Controllers/Admin/CommitteeTypeController.php
@@ -4,18 +4,24 @@ namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
 use App\Models\CommitteeType;
+use Gate;
 use Illuminate\Http\Request;
+use Symfony\Component\HttpFoundation\Response;
 
 class CommitteeTypeController extends Controller
 {
     public function index()
     {
+        abort_if(Gate::denies('committee_type_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         $committeeTypes = CommitteeType::all();
         return view('admin.committee_types.index', compact('committeeTypes'));
     }
 
     public function create()
     {
+        abort_if(Gate::denies('committee_type_create'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         return view('admin.committee_types.create');
     }
 
@@ -32,6 +38,8 @@ class CommitteeTypeController extends Controller
 
     public function edit(CommitteeType $committeeType)
     {
+        abort_if(Gate::denies('committee_type_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         return view('admin.committee_types.edit', compact('committeeType'));
     }
 
@@ -48,6 +56,8 @@ class CommitteeTypeController extends Controller
 
     public function destroy(CommitteeType $committeeType)
     {
+        abort_if(Gate::denies('committee_type_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         $committeeType->delete();
 
         return redirect()->route('admin.committee-types.index')->with('success', 'Committee Type deleted successfully.');
@@ -55,6 +65,8 @@ class CommitteeTypeController extends Controller
 
     public function massDestroy(Request $request)
     {
+        abort_if(Gate::denies('committee_type_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         CommitteeType::whereIn('id', request('ids'))->delete();
         return response(null, 204);
     }
